requirements.txt
aiohttp==3.6.2
aiomysql==0.0.20
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.7
django-cors-headers==3.2.0
django-crispy-forms==1.8.1
django-filter==2.2.0
django-formtools==2.1
django-guardian==2.1.0
django-import-export==1.2.0
django-reversion==3.0.4
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-metadata==1.3.0
isort==4.3.21
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.0.2
multidict==4.7.2
mysqlclient==1.4.6
odfpy==1.4.0
openpyxl==3.0.2
packaging==19.2
Pillow==6.2.1
pluggy==0.13.1
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.5.3
PyMySQL==0.9.2
pyparsing==2.4.5
pytest==5.3.2
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
toml==0.10.1
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.7
wcwidth==0.1.7
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
yarl==1.4.2
zipp==0.6.0
